Group file management processors
